workaround: Fix Viber's tray icons
This has been a long running issue #4 and it's been bothering me as
well.

While this "fix" does introduce holes in the sandbox, namely the entire
"org.kde.*" namespace being exposed, I think it's worth to merge it
simply due to the issues that not having an tray icon would cause. (And
the amount of frustration that brings)

Firstly, not being able to quit Viber is ONE BIG issue.
Secondly, having the ability to be aware that Viber's running on the
system regardless whether the application window is open or not is
pretty neat.
Thirdly, keeping track of background tasks via the tray icons is cool as
well.

So what's the rationale, why does Viber need an access to an entire
session bus?

Basically since Viber DOES NOT use the "org.kde.StatusNotifierItem" bus,
it instead opts to create it's own under the name
"org.kde.StatusNotifierItem-$randomNumber-1".

This has two problems:

For one: Letting Viber talk to a bus won't work as it's **initiating**
instead of using one. Hence the need for "own-name".

Secondly: Flatpak doesn't allow string regexes. It does allow
sub-package regexes, but not actual string regexes, because if that were
the case I could do "org.kde.StatusNotifierItem-*" instead.

Hopefully this gets merged so that people using Viber don't have to
scour yet another GitHub issue page and hopefully make Linux less
tedious for them.

Cheers,
nikp123